requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
nltk==3.9.1
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
google-cloud-core==2.4.1
google-auth==2.36.0
requests==2.32.3
gunicorn==23.0.0
openai==1.25.0
flask-cors==4.0.0 
markdown-it-py>=3.0.0
jsonpatch>=1.33 